+++ Graphic11/textdesigner/TextDblGlowStrategy.cpp
@@ -82,17 +82,17 @@ bool TextDblGlowStrategy::DrawString(
 {
 	using namespace Gdiplus;
 	GraphicsPath path;
-	Status status = path.AddString(pszText,wcslen(pszText),pFontFamily,fontStyle,nfontSize,ptDraw,pStrFormat);
+	Status status = path.AddString(pszText,wcslen(pszText),pFontFamily,fontStyle,(REAL)nfontSize,ptDraw,pStrFormat);
 	if(status!=Ok)
 		return false;
 
 	for(int i=m_nThickness1; i<=m_nThickness1+m_nThickness2; ++i)
 	{
-		Pen pen(m_clrOutline2,i);
+		Pen pen(m_clrOutline2,(REAL)i);
 		pen.SetLineJoin(LineJoinRound);
 		pGraphics->DrawPath(&pen, &path);
 	}
-	Pen pen1(m_clrOutline1,m_nThickness1);
+	Pen pen1(m_clrOutline1, (REAL)m_nThickness1);
 	pen1.SetLineJoin(LineJoinRound);
 	pGraphics->DrawPath(&pen1, &path);
 
@@ -121,17 +121,17 @@ bool TextDblGlowStrategy::DrawString(
 {
 	using namespace Gdiplus;
 	GraphicsPath path;
-	Status status = path.AddString(pszText,wcslen(pszText),pFontFamily,fontStyle,nfontSize,rtDraw,pStrFormat);
+	Status status = path.AddString(pszText,wcslen(pszText),pFontFamily,fontStyle, (REAL)nfontSize,rtDraw,pStrFormat);
 	if(status!=Ok)
 		return false;
 
 	for(int i=m_nThickness1; i<=m_nThickness1+m_nThickness2; ++i)
 	{
-		Pen pen(m_clrOutline2,i);
+		Pen pen(m_clrOutline2, (REAL)i);
 		pen.SetLineJoin(LineJoinRound);
 		pGraphics->DrawPath(&pen, &path);
 	}
-	Pen pen1(m_clrOutline1,m_nThickness1);
+	Pen pen1(m_clrOutline1, (REAL)m_nThickness1);
 	pen1.SetLineJoin(LineJoinRound);
 	pGraphics->DrawPath(&pen1, &path);
 
@@ -176,12 +176,12 @@ bool TextDblGlowStrategy::GdiDrawString(
 
 	for(int i=m_nThickness1; i<=m_nThickness1+m_nThickness2; ++i)
 	{
-		Pen pen(m_clrOutline2,i);
+		Pen pen(m_clrOutline2, (REAL)i);
 		pen.SetLineJoin(LineJoinRound);
 		pGraphics->DrawPath(&pen, pPath);
 	}
 
-	Pen pen1(m_clrOutline1,m_nThickness1);
+	Pen pen1(m_clrOutline1, (REAL)m_nThickness1);
 	pen1.SetLineJoin(LineJoinRound);
 	pGraphics->DrawPath(&pen1, pPath);
 
@@ -232,12 +232,12 @@ bool TextDblGlowStrategy::GdiDrawString(
 
 	for(int i=m_nThickness1; i<=m_nThickness1+m_nThickness2; ++i)
 	{
-		Pen pen(m_clrOutline2,i);
+		Pen pen(m_clrOutline2, (REAL)i);
 		pen.SetLineJoin(LineJoinRound);
 		pGraphics->DrawPath(&pen, pPath);
 	}
 
-	Pen pen1(m_clrOutline1,m_nThickness1);
+	Pen pen1(m_clrOutline1, (REAL)m_nThickness1);
 	pen1.SetLineJoin(LineJoinRound);
 	pGraphics->DrawPath(&pen1, pPath);
 
@@ -275,19 +275,19 @@ bool TextDblGlowStrategy::MeasureString(
 {
 	using namespace Gdiplus;
 	GraphicsPath path;
-	Status status = path.AddString(pszText,wcslen(pszText),pFontFamily,fontStyle,nfontSize,ptDraw,pStrFormat);
+	Status status = path.AddString(pszText,wcslen(pszText),pFontFamily,fontStyle, (REAL)nfontSize,ptDraw,pStrFormat);
 	if(status!=Ok)
 		return false;
 
-	*pfDestWidth= ptDraw.X;
-	*pfDestHeight= ptDraw.Y;
+	*pfDestWidth = (float)ptDraw.X;
+	*pfDestHeight = (float)ptDraw.Y;
 	bool b = GDIPath::MeasureGraphicsPath(pGraphics, &path, pfPixelsStartX, pfPixelsStartY, pfDestWidth, pfDestHeight);
 
 	if(false==b)
 		return false;
 
 	float pixelThick = 0.0f;
-	b = GDIPath::ConvertToPixels(pGraphics,m_nThickness1+m_nThickness2,0.0f,NULL,NULL,&pixelThick,NULL);
+	b = GDIPath::ConvertToPixels(pGraphics,(float)(m_nThickness1+m_nThickness2),0.0f,NULL,NULL,&pixelThick,NULL);
 
 	if(false==b)
 		return false;
@@ -313,19 +313,19 @@ bool TextDblGlowStrategy::MeasureString(
 {
 	using namespace Gdiplus;
 	GraphicsPath path;
-	Status status = path.AddString(pszText,wcslen(pszText),pFontFamily,fontStyle,nfontSize,rtDraw,pStrFormat);
+	Status status = path.AddString(pszText,wcslen(pszText),pFontFamily,fontStyle, (REAL)nfontSize,rtDraw,pStrFormat);
 	if(status!=Ok)
 		return false;
 
-	*pfDestWidth= rtDraw.GetLeft();
-	*pfDestHeight= rtDraw.GetTop();
+	*pfDestWidth = (float)rtDraw.GetLeft();
+	*pfDestHeight = (float)rtDraw.GetTop();
 	bool b = GDIPath::MeasureGraphicsPath(pGraphics, &path, pfPixelsStartX, pfPixelsStartY, pfDestWidth, pfDestHeight);
 
 	if(false==b)
 		return false;
 
 	float pixelThick = 0.0f;
-	b = GDIPath::ConvertToPixels(pGraphics,m_nThickness1+m_nThickness2,0.0f,NULL,NULL,&pixelThick,NULL);
+	b = GDIPath::ConvertToPixels(pGraphics,(float)(m_nThickness1+m_nThickness2),0.0f,NULL,NULL,&pixelThick,NULL);
 
 	if(false==b)
 		return false;
@@ -365,8 +365,8 @@ bool TextDblGlowStrategy::GdiMeasureString(
 		return false;
 	}
 
-	*pfDestWidth= ptDraw.X;
-	*pfDestHeight= ptDraw.Y;
+	*pfDestWidth = (float)ptDraw.X;
+	*pfDestHeight = (float)ptDraw.Y;
 	b = GDIPath::MeasureGraphicsPath(pGraphics, pPath, pfPixelsStartX, pfPixelsStartY, pfDestWidth, pfDestHeight);
 
 	if(false==b)
@@ -380,7 +380,7 @@ bool TextDblGlowStrategy::GdiMeasureString(
 	}
 
 	float pixelThick = 0.0f;
-	b = GDIPath::ConvertToPixels(pGraphics,m_nThickness1+m_nThickness2,0.0f,NULL,NULL,&pixelThick,NULL);
+	b = GDIPath::ConvertToPixels(pGraphics,(float)(m_nThickness1+m_nThickness2),0.0f,NULL,NULL,&pixelThick,NULL);
 
 	if(false==b)
 	{
@@ -432,8 +432,8 @@ bool TextDblGlowStrategy::GdiMeasureString(
 		return false;
 	}
 
-	*pfDestWidth= rtDraw.GetLeft();
-	*pfDestHeight= rtDraw.GetTop();
+	*pfDestWidth = (float)rtDraw.GetLeft();
+	*pfDestHeight = (float)rtDraw.GetTop();
 	b = GDIPath::MeasureGraphicsPath(pGraphics, pPath, pfPixelsStartX, pfPixelsStartY, pfDestWidth, pfDestHeight);
 
 	if(false==b)
@@ -447,7 +447,7 @@ bool TextDblGlowStrategy::GdiMeasureString(
 	}
 
 	float pixelThick = 0.0f;
-	b = GDIPath::ConvertToPixels(pGraphics,m_nThickness1+m_nThickness2,0.0f,NULL,NULL,&pixelThick,NULL);
+	b = GDIPath::ConvertToPixels(pGraphics,(float)(m_nThickness1+m_nThickness2),0.0f,NULL,NULL,&pixelThick,NULL);
 
 	if(false==b)
 	{
@@ -499,8 +499,8 @@ bool TextDblGlowStrategy::GdiMeasureStringRealHeight(
 		return false;
 	}
 
-	*pfDestWidth= ptDraw.X;
-	*pfDestHeight= ptDraw.Y;
+	*pfDestWidth = (float)ptDraw.X;
+	*pfDestHeight = (float)ptDraw.Y;
 	b = GDIPath::MeasureGraphicsPathRealHeight(pGraphics, pPath, pfPixelsStartX, pfPixelsStartY, pfDestWidth, pfDestHeight);
 
 	if(false==b)
@@ -514,7 +514,7 @@ bool TextDblGlowStrategy::GdiMeasureStringRealHeight(
 	}
 
 	float pixelThick = 0.0f;
-	b = GDIPath::ConvertToPixels(pGraphics,m_nThickness1+m_nThickness2,0.0f,NULL,NULL,&pixelThick,NULL);
+	b = GDIPath::ConvertToPixels(pGraphics,(float)(m_nThickness1+m_nThickness2),0.0f,NULL,NULL,&pixelThick,NULL);
 
 	if(false==b)
 	{
@@ -566,8 +566,8 @@ bool TextDblGlowStrategy::GdiMeasureStringRealHeight(
 		return false;
 	}
 
-	*pfDestWidth= rtDraw.GetLeft();
-	*pfDestHeight= rtDraw.GetTop();
+	*pfDestWidth = (float)rtDraw.GetLeft();
+	*pfDestHeight = (float)rtDraw.GetTop();
 	b = GDIPath::MeasureGraphicsPathRealHeight(pGraphics, pPath, pfPixelsStartX, pfPixelsStartY, pfDestWidth, pfDestHeight);
 
 	if(false==b)
@@ -581,7 +581,7 @@ bool TextDblGlowStrategy::GdiMeasureStringRealHeight(
 	}
 
 	float pixelThick = 0.0f;
-	b = GDIPath::ConvertToPixels(pGraphics,m_nThickness1+m_nThickness2,0.0f,NULL,NULL,&pixelThick,NULL);
+	b = GDIPath::ConvertToPixels(pGraphics, (float)(m_nThickness1+m_nThickness2),0.0f,NULL,NULL,&pixelThick,NULL);
 
 	if(false==b)
 	{
